requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
